project/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
redis==5.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
openai==1.14.0
python-dotenv==1.0.1